+++ middleman-core/fixtures/sinatra-app/config.rb
@@ -4,6 +4,9 @@ class MySinatra < Sinatra::Base
   get "/" do
     "Hello World (Sinatra)"
   end
+  get "/derp.html" do
+    "De doo"
+  end
 end
 
 map "/sinatra" do